DumpPixels and ComputePixelChecksum default to get the whole screen
Change the API arguments, make the DumpPixels and ComputePixelChecksum
methods default to get the whole screen, if no area is given.

When move to the next version of Chameleon board, getting the whole screen
is fast. We can do comparison on the whole screen instead of partial areas.

BUG=chromium:330804
TEST=manual, verified DumpPixels and ComputePixelChecksum work
>>> import xmlrpclib
>>> board = xmlrpclib.ServerProxy('http://localhost:9992')
>>> board.DumpPixels(1, 0, 0, 10, 10).data
......
>>> board.ComputePixelChecksum(1, 0, 0, 10, 10)
......

Change-Id: I9f029c433445fdf07d6dbbf3d186b3bdf7c9dbbb
Reviewed-on: https://chromium-review.googlesource.com/181990
Reviewed-by: Hung-ying Tyan <tyanh@chromium.org>
Commit-Queue: Wai-Hong Tam <waihong@chromium.org>
Tested-by: Wai-Hong Tam <waihong@chromium.org>