Scripts to deploy and start Chamleond on startup
The init.d script to start chameleond service on start-up. And the deploy
script creates the config file and installs the init.d script to the system.

BUG=chromium:345541
TEST=manaual
Ran the deploy script in Chameleon board and checked chameleond started.
Rebooted the board and checked chameleond auto-started on startup.

Change-Id: I7355428e28780d2247330af168d4ce73483f5db4
Reviewed-on: https://chromium-review.googlesource.com/186511
Reviewed-by: Hung-ying Tyan <tyanh@chromium.org>
Commit-Queue: Wai-Hong Tam <waihong@chromium.org>
Tested-by: Wai-Hong Tam <waihong@chromium.org>