Set to the default EDID on reset
Sometime the EDID on the Chameleon board may be corrupted due to a test
failure. This failure causes the Chameleon not well-functional. It is
safe to restore the EDID to a default one on reset.

BUG=chromium:347436
TEST=manual
Reset the Chameleon board in the middle of running the test display_Resolution.
As a result, the EDID was changed to others, like a different resolution one.
Started anthoer test display_HotPlugAtBoot and checked the EDID is recovered.

Change-Id: Icd97cfd1db175939fc89e1b6b4625032e2d0922e
Reviewed-on: https://chromium-review.googlesource.com/188830
Reviewed-by: Hung-ying Tyan <tyanh@chromium.org>
Tested-by: Wai-Hong Tam <waihong@chromium.org>
Commit-Queue: Wai-Hong Tam <waihong@chromium.org>