Restart HDMI receiver if FPGA cannot lock the frames
We found that the mode change register may not be set on every mode change.
It causes screen corruption during running tests.

This CL makes a work-around by checking the resolution reported by FPGA with
the one reported by HDMI receiver. If these two resolutions are not the same,
restart the HDMI receiver.

Before the above check, we should wait the video input stable. Otherwise,
the resolution we got may not be correct.

BUG=chromium:347436
TEST=test_that --arg="chameleon_host=$CHAMELEON" $DUT suite:chameleon

Change-Id: Ifdfed483c5b14e10433507e4a0cb49f6f57682f6
Reviewed-on: https://chromium-review.googlesource.com/188833
Reviewed-by: Hung-ying Tyan <tyanh@chromium.org>
Commit-Queue: Wai-Hong Tam <waihong@chromium.org>
Tested-by: Wai-Hong Tam <waihong@chromium.org>