Speed-up by extracting the selected area in the native pixeldump utility
The original approach of API DumpPixels with a selected area is:
 - Call native C program pixeldump to dump the whole screen from RAM to a file.
 - Read the file and extract the selected area in Python.

Dumping the whole screen wastes. This CL speeds up it by only dumping the small
selected area to a file. It saves ~200ms on capturing a small area.

BUG=chromium:353261
TEST=Picked several random areas. Checked the results (before and after) are
the same, by running the script:
$ screenshot --chameleon_host $CHAMELEON_IP --output image.png --area $AREA

Change-Id: I41a7f596a51767c7ecff62ca40f3a0568ffc8581
Reviewed-on: https://chromium-review.googlesource.com/190399
Reviewed-by: Wai-Hong Tam <waihong@chromium.org>
Commit-Queue: Wai-Hong Tam <waihong@chromium.org>
Tested-by: Wai-Hong Tam <waihong@chromium.org>