Mark chip error when timeout of video-stable-bit not set
The video-stable-bit on the HDMI receiver reports the ChromeBook video
stable or not. The video may be unstable during a transition like reboot or
suspend. This transition commonly is short. Mostly making the video-stable
timeout is caused by a receiver issue. So it is safer to mark it as a
chip error, such that it can be repaired. Marking it in a wrong case does
not matter, just wasting time to restarting Chameleond.

BUG=chromium:360993
TEST=Manaul
Ran Chameleon test suite on lab machines.
  $ test_that --fast $LAB_HOST suite:chameleon
Saw the log that the repair rountine was executed.

Change-Id: Ifbabc7762a9addf81c4b17c42bf7c2a800e910ba
Reviewed-on: https://chromium-review.googlesource.com/195357
Tested-by: Wai-Hong Tam <waihong@chromium.org>
Reviewed-by: Hung-ying Tyan <tyanh@chromium.org>
Commit-Queue: Wai-Hong Tam <waihong@chromium.org>