Chameleon: Implement FireHpdPulse() for the fpga_tio driver.
The original hpd_control.c is renamed to hpd_control_hdmi.c as it is for the
fpga_hdmi driver. New hpd_control_tio.c is added. It is based on
hpd_control_hdmi with the addition of new arguments for specifying io addresses
for different video connectors.

BUG=chromium:356064
TEST=by the manually run the test on peppy and see it passed.

Change-Id: I2aad740516165b76d9db0e07ac9c433676a27b39
Reviewed-on: https://chromium-review.googlesource.com/208290
Reviewed-by: Hung-ying Tyan <tyanh@chromium.org>
Commit-Queue: Ting-Yuan Cheng <tingyuan@chromium.org>
Tested-by: Ting-Yuan Cheng <tingyuan@chromium.org>