Add VGA cable detection
This change implements the detection by checking the "Sync Detect" register.
If the source does not produce video signal, then it will not work. However,
this should not be a concern as Chameleon tests are run in a controlled
environment where the above scenario would not happen.

BUG=chromium:374103
TEST=Manually tested it in Python:
>>> import xmlrpclib
>>> s = xmlrpclib.ServerProxy('http://10.10.10.15:9992')
>>> s.ProbeInputs()
[4]
>>> s.GetConnectorType(4)
'VGA'
>>> s.IsPhysicalPlugged(4)
False
(plug VGA cable)
>>> s.IsPhysicalPlugged(4)
True

Change-Id: Idb5a778d8fc55a81bd8dc6505e288b966d349d28
Reviewed-on: https://chromium-review.googlesource.com/211986
Reviewed-by: Wai-Hong Tam <waihong@chromium.org>
Commit-Queue: Wai-Hong Tam <waihong@chromium.org>
Tested-by: Wai-Hong Tam <waihong@chromium.org>