Hint FPGA to use a full-screen capture if the area is not given
FPGA supports a full-screen capture and a cropped-screen capture. There
is an alignment requirement (8-byte/16-byte) for the cropped-screen capture.
If the area is not given via the API arguments, should use a full-screen
capture, instead of a cropped screen.

We don't need a resolution like 1440x900 failed due to the unaligned
screen size.

BUG=chromium:385252
TEST=Ran the display_Resolution test passed.

Change-Id: I53d4d6196c3b2b77c1cccd24f46dce81786e89a3
Reviewed-on: https://chromium-review.googlesource.com/213351
Reviewed-by: Hung-ying Tyan <tyanh@chromium.org>
Commit-Queue: Wai-Hong Tam <waihong@chromium.org>
Tested-by: Wai-Hong Tam <waihong@chromium.org>