Add VGA plug and unplug logic
VGA has no HPD to detect hot-plug. By default, the signal is always
connected. There is a hardware on Chameleon board to block the RGB
source signal to emulate disconnection. We use this block logic to
emulate plug and unplug.

The cable check logic tries to plug and wait for the H-Sync/V-Sync
signal recevied from the source. It does not work if DUT is not well-
behaved.

BUG=chromium:429931
TEST=Manually tested by the APIs ProbeInputs, Plug, Unplug, and IsPlugged.

Change-Id: I7a6ca044d46074ecd2f09bf6f0e245534521ff84
Reviewed-on: https://chromium-review.googlesource.com/227541
Reviewed-by: Wai-Hong Tam <waihong@chromium.org>
Tested-by: Wai-Hong Tam <waihong@chromium.org>
Commit-Queue: Wai-Hong Tam <waihong@chromium.org>