Support setting VGA mode
This change exposes a new API to set the proper mode for VGA. A mode is
like 'PC_1920x1080x60'. Setting the mode triggers the receiver to operate
properly.

Before we have an automatical way to detect all the modes, this API solves
it by moving the job to the test side. A test is relatively easy to check
the current DUT resolution and set the mode.

BUG=chromium:429931
TEST=Manually switched different mode and called the API.

Change-Id: I4bb484ec56131120a5318c8e6e68836c8e011789
Reviewed-on: https://chromium-review.googlesource.com/227786
Tested-by: Wai-Hong Tam <waihong@chromium.org>
Reviewed-by: Hung-ying Tyan <tyanh@chromium.org>
Commit-Queue: Wai-Hong Tam <waihong@chromium.org>