Support VGA mode auto-detection
Add the support of auto detecting the VGA mode by probing the counter of
HSync. This detection logic is simplified and only works on 60Hz reflresh
rate.

By iterating all the resolutions on Parrot, the results are:
 - worked on 720x480, 800x600, 1280x720, 1360x768, 1280x800, 1280x960,
             1680x1050, 1280x1024, 1920x1080, 1600x1200, 1920x1200
 - failed on 720x400, 640x480, 1440x480, 832x624, 1024x768, 1152x864, 1440x900

The failed cases seems to be problems of the driver or the receiver. It
needs to be studied more.

The solution is to set the VGA mode properly and skip testing the totally-
not-supported resolutions.

BUG=chromium:429931
TEST=Tested on Parrot, by switching all the supported resolutions on extended
mode. Work fine except the resolutions stated above.

Change-Id: I9fc61eab2f22f40376a4575adbc18456fbbbb4ee
Reviewed-on: https://chromium-review.googlesource.com/229189
Reviewed-by: Hung-ying Tyan <tyanh@chromium.org>
Commit-Queue: Wai-Hong Tam <waihong@chromium.org>
Tested-by: Wai-Hong Tam <waihong@chromium.org>