Set the correct dependency in Makefile
The chameleond distribution should depend on the binaries. This dependency
should be set explicitly. Otherwise, the parallel-make may create the
chameleond distribution in the same time of compiling the binaries. As a
result, some zero-sized binaries (in the progress of compiling) may be used.

BUG=chromium:435186
TEST=Ran 'make -j40' to see compiling binaries before making distribution.

Change-Id: I19d7117853bb2af6b9c4324bfe54ba85b789da5f
Reviewed-on: https://chromium-review.googlesource.com/231073
Tested-by: Wai-Hong Tam <waihong@chromium.org>
Reviewed-by: Hung-ying Tyan <tyanh@chromium.org>
Commit-Queue: Wai-Hong Tam <waihong@chromium.org>