Reset receiver before waiting for the input signal stable
We saw a case that when the EDID changes. The input signal is not stable
after waiting quite a long time. It may a wrong report from the receiver.
Better to handle the interrupts first and reset the receiver if needed.
So move the reset logic at the beginning.

It also explicitly raises the exception to avoid some hard-to-explain
problems after that.

BUG=chromium:435708
TEST=Ran Chameleon tests and passed on nyan_big.

Change-Id: I45e67830baca8239a14a203162fb8ffbf027f798
Reviewed-on: https://chromium-review.googlesource.com/231439
Tested-by: Wai-Hong Tam <waihong@chromium.org>
Reviewed-by: Hung-ying Tyan <tyanh@chromium.org>
Commit-Queue: Wai-Hong Tam <waihong@chromium.org>