Add an accurate way to stop capturing video
We used to stop capturing video by monitoring the current frame index
reaches a predefined one, like 5 seconds since the start of the video.
When calling the StopCapturingVideo API, it may really stop several
frames later due to the network latency.

Now we introduce an accurate way to stop by passing a stop_index to
the StopCapturingVideo API. As the monitor logic is executed on the
Chameleon board, it stops accurately at the given index.

A test can use the monitor logic in CL:235531 to find the start_index
and use the following code to stop, like 5 seconds later.

  end_index = start_index + 5 * 60  # capture 5 seconds in 60fps
  c.StopCapturingVideo(end_index)

BUG=chromium:385252
TEST=Played a video on DUT and executed the above code to test.

Change-Id: I47cf49b926903511a98bc93a1f9c29b5a7011b6f
Reviewed-on: https://chromium-review.googlesource.com/236456
Tested-by: Wai-Hong Tam <waihong@chromium.org>
Reviewed-by: Mussa Kiroga <mussa@chromium.org>
Reviewed-by: Hung-ying Tyan <tyanh@chromium.org>
Commit-Queue: Wai-Hong Tam <waihong@chromium.org>