Set pixel mode based on pixel clock.
The firmware for the 6803 reference board sets the rx in dual pixel
mode only when the pixel clock is greater than 160. This change does
similar adjustment in InputFlow.Do_FSM(). The threshold we use is 130
instead of 160 as we found that the FPGA works more reliably when
pixel clock is under this value.

BUG=chromium:443080
TEST=Ran suite:chameleon_hdmi on nyan_big, squawks, peppy and link
multiple times and observed the tests expected to pass passed reliably.

Change-Id: Iac7140ccffa83af1a313e88d07ac7dbf3b4a81bd
Reviewed-on: https://chromium-review.googlesource.com/238143
Tested-by: Hung-ying Tyan <tyanh@chromium.org>
Reviewed-by: Hung-ying Tyan <tyanh@chromium.org>
Commit-Queue: Hung-ying Tyan <tyanh@chromium.org>