Change variable names which process fields from 'frame' to 'field'
The signal which outputs from receivers is actually on a field basis,
not a frame basis. It works fine in progressive mode as one field is
a full frame. But in interlaced mode, one field is only half of a
frame.

This change is to reflect the naming of it. The original FrameManager
is actaully a FieldManager. A real FrameManager will be added later.

BUG=chromium:468100
TEST=Ran suite:chameleon_hdmi still passed.

Change-Id: I5321a7d7ad6640c6e252f131e5be3b71711726e8
Reviewed-on: https://chromium-review.googlesource.com/260998
Trybot-Ready: Wai-Hong Tam <waihong@chromium.org>
Tested-by: Wai-Hong Tam <waihong@chromium.org>
Reviewed-by: Hung-ying Tyan <tyanh@chromium.org>
Commit-Queue: Wai-Hong Tam <waihong@chromium.org>