Set pixel mode based on pixel clock for DP
Single pixel mode cannot handle 2560x1600. This commit changes rx to
dual pixel mode when pixel clock is greater than or equal to 200MHz.

Note: It is found that pixel clock reading is not reliable. For
example, a value of 251MHz was read for 640x480. Before we figure out
what is going on there, a hack is added in this commit to enforce
single pixel mode when the frame width is below 1920.

BUG=chromium:429928
TEST=successfully ran display_EdidStress and display_ResolutionList
on link over DP

Change-Id: I2f6817a15f5cfae60452e7a18807c398bc98e603
Reviewed-on: https://chromium-review.googlesource.com/262580
Tested-by: Hung-ying Tyan <tyanh@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@chromium.org>
Commit-Queue: Hung-ying Tyan <tyanh@chromium.org>