audio_board: add _JackPlugger to control plugger
Add _JackPlugger to provide interface to control plugger.
The plugger is a motor which can plug/unplug a 4-ring 3.5mm audio cable
to/from audio jack of Cros device. This motor is controlled by 4 pins on
audio board on an I/O expander.

Initialize a _JackPlugger in AudioBoard. If audio board fails to
initialize jack plugger, assume this configuration does not have audio
box. Raises exception if SetJackPlugger of AudioBox is called in this
case.

BUG=chromium:472906
TEST=Check plugger is disabled when chameleond resets.
TEST=Check chameleond starts fine without audio box.

Change-Id: Ibb5a4c0e5b291d52f066a5a44fe40f40b9adf668
Reviewed-on: https://chromium-review.googlesource.com/263675
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>
Commit-Queue: Cheng-Yi Chiang <cychiang@chromium.org>
Tested-by: Cheng-Yi Chiang <cychiang@chromium.org>