Adjust the pixel mode threshold values for HDMI RX
Recently we found that the HDMI RX output becomes unstable when it
operates in dual pixel mode for the input signal with its pixel clock
slightly greater than 125MHz. The detail of the result is reported in
http://crbug.com/453416. This commit changes the pixel mode threshold
values from [115, 125] to [126, 130] to ensure that the RX operates in
single pixel mode when the pixel clock is less than 126MHz.

BUG=chromium:453416
TEST=Passed display_ResolutionList and display_Resolution on nyan_blaze.

Change-Id: Ic306ecfdccc9c1da9f9a298c0c05d377373ef36d
Reviewed-on: https://chromium-review.googlesource.com/270208
Tested-by: Hung-ying Tyan <tyanh@chromium.org>
Reviewed-by: Yuli Huang <yuli@chromium.org>
Commit-Queue: Hung-ying Tyan <tyanh@chromium.org>