USBController: Refactor to encapsulate USB configs in use by driver
Two private instance variables are added:
1)_driver_configs_to_set is the USBAudioDriverConfigs object used to
store configurations set by the user through SetDriverPlaybackConfigs
and SetDriverCaptureConfigs methods. It is initially set to the
default values of a USBAudioDriverConfigs object at instantiation of
the USBController.
2)_driver_configs_in_use is the USBAudioDriverConfigs object
representing the set of configurations currently in use by the USB audio
driver when it is enabled/modprobed. It should be set to None when the
driver is not enabled.
These two variables are used to encapsulate driver configurations such
that driver_configs_in_use represents the true set of configurations
used by the driver whenever it is enabled, and such that getter methods
for supported data formats will always return the correct format
supported by the driver at the moment.

BUG=chromium:514541
TEST=not used yet.

Change-Id: I9ac28bde6964f3bf942b649fc6819e8d2a712371
Reviewed-on: https://chromium-review.googlesource.com/295320
Commit-Ready: Cheng-Yi Chiang <cychiang@chromium.org>
Tested-by: Cheng-Yi Chiang <cychiang@chromium.org>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>