USBController: Refactor and add _CheckModprobeResultAndUpdateConfigs
The new method determines the result of the modprobe command to insert
the USB audio driver module into the kernel, based on the command line
output from calling the command. Based on the result, the method also
updates driver_configs_in_use, which is used to represent the set of
configurations that is modprobed together with the driver.

BUG=chromium:514541
TEST=not used yet.

Change-Id: I452b2b6da6ecb525d747adfccbef2f234f66c7fb
Reviewed-on: https://chromium-review.googlesource.com/295326
Commit-Ready: Cheng-Yi Chiang <cychiang@chromium.org>
Tested-by: Cheng-Yi Chiang <cychiang@chromium.org>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>