usb_configs: Explain how 'file_type' attribute is handled here
USBAudioDriverConfigs class saves configurations in AudioDataFormat
objects, which has the 'file_type' attribute. However, 'file_type' is
not a relevant parameter for USB audio driver configurations. Therefore,
it is ignored in this class. AudioDataFormat objects are used in this
class because, in spite of 'file_type', they organize and encapsulate
driver configurations rather well.

BUG=chromium:514541
TEST=not used yet.

Change-Id: I4669844571cfda5d9ee6de9ed96c6b37ceec768a
Reviewed-on: https://chromium-review.googlesource.com/296862
Commit-Ready: Cheng-Yi Chiang <cychiang@chromium.org>
Tested-by: Cheng-Yi Chiang <cychiang@chromium.org>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>