Add bluetooth_rn42.py to support RN-42 bluetooth evaluation kit
The purpose is to use the kit to emulate bluetooth peripherals such as
keyboard, mouse, etc. on Chameleon to build a bluetooth test framework
for HID profile. As a first step, this patch supports the serial
communication capability between a host and the chip as well as the
basic configuration of the kit.

BUG=chromium:575551
TEST=Steps as below
1. Connect a set of RN-42 kit to a chromebook or a Chameleon board
   with host-mode USB port.
2. Run "python bluetooth_rn42.py".
3. It would output something like
   enter: True
   chip name: RNBT-A955
   firmware version: Ver 6.15 04/26/2013 (c) Roving Networks
   operation_mode: MASTER
   authentication_mode: PIN_CODE
   service_profile: HID
   local_bluetooth_mac: 00066675A955
   connection_status: False
   remote_bluetooth_mac: None
   HID_device_class: KEYBOARD
   leave: True

Change-Id: Idc9153acd3b1a3699914b173ef63fb1151c54b93
Reviewed-on: https://chromium-review.googlesource.com/322061
Commit-Ready: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>