Add bluetooth_hid module to emulate a keyboard
This module provides an abstraction of blueooth HID devices.
As a first step, it emulates a keyboard, connects to a dut,
and sends some texts to the dut.

Note that the test below is only a simple demo of using this module.
More commonly, the pairing and connection requests are initiated by
a host instead of a peripheral device. That requires some host side
programming that would be provided later.

BUG=chromium:575551
TEST=The RN-42 emulates a keyboard and continuously sends
some texts to a dut.
1. Plug in an RN-42 kit to a chromebook or a Chameleon board
   with host-mode USB port.
2. Assume the bluetooth address of the dut is 00:12:34:56:78:90
   Open a virtual terminal in a chrome broswer.
3. Run the following command:
   python bluetooth_hid.py 00:12:34:56:78:90 "echo hello world"
4. Observe the following output:
   Creating an emulated bluetooth keyboard...
   Connecting to the remote address 00:12:34:56:78:90
   Sending "echo hello world" for the 1th time...
   Sending "echo hello world" for the 2th time...
   Sending "echo hello world" for the 3th time...
   ...
   Sending "echo hello world" for the 10th time...
   Disconnecting...
   Closing the keyboard...

   The "echo hello world" message should show up in the
   virtual terminal without any characters getting lost.

Change-Id: Ia28ed7c890db94ea066998d999af44676bfab9f8
Reviewed-on: https://chromium-review.googlesource.com/322942
Commit-Ready: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@chromium.org>