[Chameleon] refactor USB module controller
In previous usb.py USBController is dedicated for audio module only.
Since we are going to have more USB modules supported (ex. HID),
USBController needs to be refactored to be generalized enough for all
modules, while USBAudioController is the sub-class to USBController to
have audio-related functions.

usb_configs.py is renamed as usb_audio_configs because it is also
dedicated for audio only.

BUG=None
TEST=manual test

Change-Id: Ice3c5d767ed8b9ec5860e0e49cc09c67d8bd4693
Reviewed-on: https://chromium-review.googlesource.com/330106
Commit-Ready: Pin-chih Lin <johnylin@chromium.org>
Tested-by: Pin-chih Lin <johnylin@chromium.org>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>