Implement the utility to compute histograms of frames
This is a C program running on Chameleon board to compute the
histograms of sampled pixels from the Chameleon framebuffer.

A user should give the start addresses (support multiple)
of the framebuffer, the width and height of the screen.
Two configurable parameters, grid_num and sample_num,
makes it flexible for different scenarios.

The histograms are computed per 3 colors R, G, and B, using
4 buckets, i.e. values 0~63, 64~127, 128~191, and 192~255.

BUG=chromium:598387
TEST=Ran the histogram program
$ histogram 960 1080 -a 3221225472 -s 10 -g 3
$ histogram 960 1080 -a 3221225472 -a 3224338432 -s 5 -g 5

Change-Id: I4127cd054dfef2cb82999260fdf01a44da4c9ef2
Reviewed-on: https://chromium-review.googlesource.com/335334
Commit-Ready: Wai-Hong Tam <waihong@chromium.org>
Tested-by: Wai-Hong Tam <waihong@chromium.org>
Reviewed-by: Hsu Wei-Cheng <mojahsu@chromium.org>