Fix an issue of executing "make remote-install" multiple time
This issue was introduced after a change touching all files to
reset their modification time.

On the first "make remote-install", the python code is built under
the directory /tmp/$(BUNDLEDIR)/build. On the next time running
"make remote-install", all directories are overwritten except the
build directory. By touching all file, the modification time of the
previous build directory is the same as the new python code, so
setup.py thinks it is good to use, as a result, the old code is
used.

A fix is to remove the entire previous bundle directory.

BUG=None
TEST=Ran "make remote-install" multiple time with different code.

Change-Id: Iec8a9302640c1e63603440c144798ae234c22e6c
Reviewed-on: https://chromium-review.googlesource.com/332252
Commit-Ready: Wai-Hong Tam <waihong@chromium.org>
Tested-by: Wai-Hong Tam <waihong@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>