updatable/fpga_tio: enable ftdi_sio and usbserial modules in kernel config
For the chameleon board to be able to control RN-42 emulation
kit through the serial, i.e., port /dev/ttyUSB0, it is required to
mark both CONFIG_USB_SERIAL and CONFIG_USB_SERIAL_FTDI_SIO as
loadable kernel modules.

This patch updates the kernel to support the above two modules.
Refer to Gerrit for the correspond patch on the kernel config:
https://chromium-review.googlesource.com/#/c/343395/

BUG=chromium:610238
TEST=Use the commands
$ modprobe ftdi_sio
$ lsmod
Change-Id: Ib71fbfb48c513f51a84616e83c8622faf4606aed
Reviewed-on: https://chromium-review.googlesource.com/343477
Commit-Ready: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@chromium.org>