Resubmission: Control bluetooth hid devices on Chameleon board
Resubmission of the patch since pyserial package could
be installed on a chameleon board from chroot now.
This patch was reviewed on Reviewed-on:
https://chromium-review.googlesource.com/352400

This patch supports to control the bluetooth mouse emulated
by an RN-42 kit. As we have already had bluetooth HID mouse
object to control the kit, it would be cleaner to instantiate
a bluetooth_mouse object in the fpga_tio driver. The
bluetooth_hid_flow module is implemented for the control
interface of the bluetooth mouse.

BUG=chromium:610238
TEST=Use the commands

Change-Id: I73367c9dbfe832999c5918878a76ab09c57a8c15
Reviewed-on: https://chromium-review.googlesource.com/353494
Commit-Ready: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@google.com>