Control bluetooth hid devices on Chameleon board
This patch supports to control the bluetooth mouse emulated
by an RN-42 kit. As we have already had bluetooth HID mouse
object to control the kit, it would be cleaner to instantiate
a bluetooth_mouse object in the fpga_tio driver. The
bluetooth_hid_flow module is implemented for the control
interface of the bluetooth mouse.

BUG=chromium:610238
TEST=Use the commands

Change-Id: I4a6c5906dd2c7e928c7f7d7749de957b74ae0e03
Reviewed-on: https://chromium-review.googlesource.com/343479
Commit-Ready: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@chromium.org>