[Chameleon] show IP address on the last line of LCM display
For the usage of LCM display, showing IP address is probably the most
common feature for user without requesting serial console. So I put IP
address on the last line and is always shown in menu state. Moreover,
the key instruction is also shown on the same line as the guidance of
respective direction for four keys.

PageRender() is also enhanced to show a "please wait..." screen when the
fuction is executed but not responded yet, so user won't be confused if
the fuction is executed or not.

BUG=None
TEST=make remote-install to Chameleon board and see the LCM display.

Change-Id: I669278184d1282c98b1dbd64ca8941c599ea961c
Reviewed-on: https://chromium-review.googlesource.com/342487
Commit-Ready: Pin-chih Lin <johnylin@chromium.org>
Tested-by: Pin-chih Lin <johnylin@chromium.org>
Reviewed-by: Tom Tam <waihong@google.com>