Connect to the serial device only when needed
If the USB serial cable is not connected, BluetoothHIDMouseFlow
would still try to connect to the serial device and incur an
exception. This patch fixes the bug by connecting to the
serial device only when the serial device is actually needed
in an autotest.

BUG=chromium:610238
TEST=unplug the USB serial cable and observe that
chameleond could be started successfully.

Change-Id: I7d6baab577290792562a7d26cbc9ee85a0bc0586
Reviewed-on: https://chromium-review.googlesource.com/350633
Commit-Ready: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@google.com>