Install pip and pyserial
The pyserial package is required to connect to a serial device.
It is desirable to install pip which in turn could install pyserial.

BUG=chromium:620278
TEST=make and install as follows
(cr) $ make
(cr) $ make remote-install CHAMELEON_HOST=$CHAMELEON_IP
And observed that pip and pyserial have been installed on chameleon board.

Change-Id: I67ad8fe98ac9ed158efbcf37e12756844088be60
Reviewed-on: https://chromium-review.googlesource.com/352940
Commit-Ready: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@google.com>