bluetooth_rn42/hid: return True or non-None object
Every public member method of bluetooth_rn42 and bluetooth_hid
classes should return True or non-None object if the execution is
successful, and return False or raising an exception otherwise.
In this way, the caller is easier to determine whether the
method succeeds or not.

BUG=chromium:610238
TEST=None

Change-Id: I5d6a195ac33f41f4b705d4d5b87d850ab96eb044
Reviewed-on: https://chromium-review.googlesource.com/357683
Commit-Ready: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@google.com>