Set correct time
If the clock skew on a chameleon board varies too much,
pip installation pip may fail. This patch sets correct time
before deploying pip.

Note that the tools on a chameleon board are rather limited.
It is not possible to use ntpd to sync time. The date command
for taking time format is also rather restrictive. It seems
that syncing time with the host is a more feasible approach.

BUG=chromium:624487
TEST=Perform the following steps
(cr) $ make
(cr) $ make remote-install CHAMELEON_HOST=$CHAMELEON_IP

And observed that the time at the chameleon board is
reasonably precise.

Change-Id: Ib1b12d0a379099abf6dab778609a0a638542e7a7
Reviewed-on: https://chromium-review.googlesource.com/357681
Commit-Ready: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>