Synchronize time with the chameleon mirror server
This patch provides a simple way to query the time on the
chameleon mirror server and adjust the time on the chameleon
board when running chameleon-updater. Since the date/time
utilities on the bundle are rather limited, we need to do
the date/time conversion by ourselves.

Note that this patch carefully avoids touching chameleon-updater
because one would like to use the old chameleon-updater to
refresh the chameleon bundle and the time.

BUG=chromium:624487
TEST=After this patch gets merged, perform the steps on a chameleon
$ /etc/init.d/chameleon-updater start
And observe that it synchronizes the time with the chameleon mirror
server and pip is installed correctly.

If running from a host, it would simply adjust time from the host.
(cr) $ make
(cr) $ make remote-install CHAMELEON_HOST=$CHAMELEON_IP

Change-Id: I0b9885bbe27025b4688f8967dd921ed4f46ac25c
Reviewed-on: https://chromium-review.googlesource.com/359190
Commit-Ready: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>