chameleon-updater: add -O option
Add the "-O" option for wget so that the same bundle downloaded
to a chameleon board would be overwritten. Otherwise, ".1", ".2"
would be appended to the bundle name which is not desired.

BUG=chromium:624487
TEST=Run "/etc/init.d/chameleon-updater start" repeatedly on
a chameleon and observe that there is only one latest bundle
under /tmp.

Change-Id: I848473ef5ed7b78efc58ecf9aac592aaecd50a1e
Reviewed-on: https://chromium-review.googlesource.com/361911
Commit-Ready: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>