Remove dummy FireHpdPulse and FireMixedHpdPulses for VGA to use fallback
This removes the dummy FireHpdPulse and FireMixedHpdPulses routines
for the VGA input flow, allowing to use generic fallback
implementations that use calls to Plug/Unplug.

These can be quite handy for VGA as well, even though their maximum
frequency is probably more limited. This is because hot-plug detection
is done through RGB lines load detection instead of a dedicated line.

BUG=None
TEST=Send some pulses for VGA and ensure they are detected correctly

Change-Id: I5aff07819eca5b66c3c721b6c0d47e4224e951ed
Signed-off-by: Paul Kocialkowski <paul.kocialkowski@linux.intel.com>
Reviewed-on: https://chromium-review.googlesource.com/539916
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>