Modularize Bluetooth peripheral emulation kits
A BluetoothHID instance will now instantiate a particular kit
implementaton, take ownership of the instance, and redirect all
unhandled attribue accesses onto it. This means that no changes are
necessary to most clients, as in most cases this is indistinguishable
from the kit being the superclass of BluetoothHID. For those cases that
were not handled, notably accessing kit constants as though they were
part of the BluetoothHID class, begin the process of moving them to
the more general class, by redirecting them in BluetoothHID.

The RN42 module still defines part of the interface to the kit, exposed
over Chameleon-Autotest RPC. This approach allows us to swap out the
implementation defined in RN42 for one that uses a different kit, but
deliberately does not handle factoring out code common to all kits.
(The __del__ was removed in BluetoothHID because the kit should clean up
after itself.)

The flow that a client connects to will choose a particular class
that implements the required methods to be a conformant
peripheral emulation kit implementation. (Currently, that means
matching the methods and functionality of the existing RN42 class.)

Ideally, we would refactor the code so that BluetoothHID exposes an
interface to Autotest that is both kit-independent and less specifically
targeted at the Bluetooth BR/EDR case. This will require significant
modifications to the tests, hence it outside the scope of this bug.
Specifically, the refactor should expose methods in BluetoothHID* only,
ensuring that the interface between the kit and BluetoothHID* is
private, so that clients see only a generic Bluetooth HID with a
well-defined interface, with support for checking BR/EDR-specific vs
LE-specific properties.

BUG=chromium:752719
TEST=Run this code on the Chameleon board:
$ make && make remote-install CHAMELEON_HOST=$CHAMELEON_IP
Execute the non-flaky non-stress tests that use this code,
see that the tests pass:
$ test_that --board ${BOARD} --args "chameleon_host=${CHAMELEON_IP}" \
${DUT_IP} bluetooth_AdapterPairing.mouse \
bluetooth_AdapterPairing.mouse.pairing_twice \
bluetooth_AdapterHIDReports.mouse

Change-Id: Id37d960699aa8364654eb9f90e10987a1a3e7c74
Reviewed-on: https://chromium-review.googlesource.com/611251
Commit-Ready: Alexander Lent <alent@google.com>
Tested-by: Alexander Lent <alent@google.com>
Reviewed-by: Wai-Hong Tam <waihong@google.com>
Reviewed-by: Alexander Lent <alent@google.com>