Remove unused Disconnect method in BluetoothHIDFlow
This method does not appear to be used, and has a name collision
with code in the RN42 class, which we do use.

Flip the method resolution order in BluetoothHIDMouseFlow,
so that if this conflict happens again, we'll be more likely
to catch it when editing the BluetoothHIDMouse hierarchy.

Added a note to help remind future readers that this functionality
exists in case it becomes relevant, as other flows seem to use it.

BUG=chromium:752719
TEST=Run this code on the Chameleon board:
$ make && make remote-install CHAMELEON_HOST=$CHAMELEON_IP
Execute the non-flaky non-stress tests that use this code,
and watch them pass:
$ test_that --board ${BOARD} --args "chameleon_host=${CHAMELEON_IP}" \
${DUT_IP} bluetooth_AdapterPairing.mouse \
bluetooth_AdapterPairing.mouse.pairing_twice \
bluetooth_AdapterHIDReports.mouse

Change-Id: Id417203ebbda06178466211bf95e8720479fe0b1
Reviewed-on: https://chromium-review.googlesource.com/609684
Commit-Ready: Alexander Lent <alent@google.com>
Tested-by: Alexander Lent <alent@google.com>
Reviewed-by: Wai-Hong Tam <waihong@google.com>
Reviewed-by: Miao-chen Chou <mcchou@chromium.org>