deploy: Pin setuptools at 24.0.2
The new setuptools (37.0.0) expects importlib to be installed. This is
not true on the Chameleon and causes a failure during make install. This
is a hacky fix to pin the version of setuptools to 24.0.2 until a more
permanent (potentially upstream) fix can be found.

BUG=b:69849509
TEST=Setup a Chameleon from scratch without errors

Change-Id: I4700f5edd063b0cfdb69e642277f79468b737f45
Reviewed-on: https://chromium-review.googlesource.com/794515
Commit-Ready: Alec Thilenius <athilenius@chromium.org>
Tested-by: Alec Thilenius <athilenius@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@google.com>